Refactoring: utility functions moved into its own module
